Support context in publish and subscribe
Added context for the method of sending and subscribing to events, as
well as in command handlers. This is necessary to solve the problem
with the obvious inability to pass the context.

The context embedding option is not considered. Do not store Contexts
inside a struct type; instead, pass a Context explicitly to each
function that needs it. The Context should be the first parameter,
typically named ctx.

https://golang.org/pkg/context